suppot msResolution

+++ src/test/java/com/aliyun/hitsdb/client/TestHiTSDBClientQueryStringValue.java
@@ -54,8 +54,8 @@ public class TestHiTSDBClientQueryStringValue {
         
         List<QueryResult> result = tsdb.query(query);
         QueryResult queryResult = result.get(0);
-        LinkedHashMap<Integer, String> sdps = queryResult.getSdps();
-        LinkedHashMap<Integer, Number> dps = queryResult.getDps();
+        LinkedHashMap<Long, String> sdps = queryResult.getSdps();
+        LinkedHashMap<Long, Number> dps = queryResult.getDps();
         System.out.println(sdps);
         System.out.println(dps);
         System.out.println("查询结果：" + result);
